+++ app/src/pages/manage/competencies/[id]/create.vue
@@ -1,16 +1,6 @@
 <script lang="ts" setup>
-import { FormKitNode } from '@formkit/core';
-
-import { useDemoStore } from '~/demo-store';
-import {
-  GetCompetencyQuery,
-  CreateCompetencyMutation,
-  CreateCompetencyMutationVariables,
-} from '~/generated/graphql';
-import { CREATE_COMPETENCY_QUERY, READ_COMPETENCY_QUERY } from '~/gql';
-
-const demoStore = useDemoStore();
-const router = useRouter();
+import { GetCompetencyQuery } from '~/generated/graphql';
+import { READ_COMPETENCY_QUERY } from '~/gql';
 
 const props = defineProps<{
   id: string; // route param
@@ -26,48 +16,6 @@ const {
   { fetchPolicy: 'network-only' },
 );
 const competency = useResult(result);
-
-const { mutate: createCompetency } = useMutation<
-  CreateCompetencyMutation,
-  CreateCompetencyMutationVariables
->(CREATE_COMPETENCY_QUERY);
-
-const formErrors = ref<string[]>([]);
-const formNode = ref<FormKitNode>();
-const formValues = ref<{ title: string }>({ title: '' });
-
-async function handleFormSubmit(): Promise<void> {
-  formErrors.value = [];
-  if (demoStore.activeUserId == null) {
-    formErrors.value.push('No active user id selected.');
-    return;
-  }
-  try {
-    const response = await createCompetency({
-      currentUserId: demoStore.activeUserId,
-      data: { parentId: props.id, title: formValues.value.title },
-    });
-    switch (response?.data?.createCompetency.__typename) {
-      default:
-        throw new Error('unknown api response');
-      case 'CreateCompetencyErrorPayload': {
-        const { error: createError } = response.data.createCompetency;
-        const fieldNode = formNode.value?.at(createError.path);
-        if (fieldNode) {
-          fieldNode.setErrors([createError.message]);
-        } else {
-          formErrors.value.push(createError.message);
-        }
-        break;
-      }
-      case 'CreateCompetencySuccessPayload':
-        router.push(`/manage/competencies/${props.id}`);
-        break;
-    }
-  } catch {
-    formErrors.value.push('Something went wrong');
-  }
-}
 </script>
 <template>
   <template v-if="readError">
@@ -83,16 +31,6 @@ async function handleFormSubmit(): Promise<void> {
     <ui-backbutton :to="`/manage/competencies/${props.id}`">
       {{ competency.title }}
     </ui-backbutton>
-    <h2 class="text-xl mb-3">Create competency</h2>
-    <FormKit
-      v-model="formValues"
-      type="form"
-      submit-label="Create competency"
-      :errors="formErrors"
-      @node="formNode = $event"
-      @submit="handleFormSubmit"
-    >
-      <FormKit name="title" label="Title" type="text" validation="required" />
-    </FormKit>
+    <create-competency :id="`${props.id}`"></create-competency>
   </template>
 </template>
